Fixed getch() bug
!!USE getch only after refresh!!

+++ src/Tiles.cpp
@@ -1,6 +1,5 @@
 #include "Tiles.hpp"
 
-#include <ncurses.h>
 
 void WhiteTile::drawSelf(WINDOW* wnd, const Pos& pos, const Pos& size) {
 	static int format = Glob::instance().GetColorAttrib(COLOR_BLACK, COLOR_WHITE); 
@@ -29,7 +28,7 @@ void BlackTile::drawSelf(WINDOW* wnd, const Pos& pos, const Pos& size) {
 LabelTile::LabelTile(std::string_view msg) : m_msg(std::move(msg)){}
 
 void LabelTile::drawSelf(WINDOW* wnd, const Pos& pos, const Pos& size) {
-	const int begin = -m_msg.size() / 2, end = begin + m_msg.size();
+	const int begin = -static_cast<int>(m_msg.size()) / 2, end = begin + m_msg.size();
 	int wp = -(size[0] * size[1] / 2) - 1;
 	for(int y = 0; y < size[1]; ++y) {
 		wmove(wnd, pos[1]  + y, pos[0]);
